issue #155 - move standalone console tool into Bridge/Standalone subfolder
