+++ src/Bridge/Symfony/DependencyInjection/DbToolsConfiguration.php
@@ -17,6 +17,23 @@ class DbToolsConfiguration implements ConfigurationInterface
         return '%kernel.project_dir%/var/db_tools';
     }
 
+    /**
+     * Append values in configuration we cannot set a default.
+     *
+     * For example, 'anonymizer_paths', if set by the user, will loose the
+     * default anonymizer paths, and we need them to be set in all cases.
+     *
+     * So we act after configuration has been processed and restore missing
+     * values from here. This also allows the standalone configuration doing
+     * it outside of Symfony extension context.
+     */
+    public static function appendPostConfig(array $config): array
+    {
+        $config['anonymizer_paths'][] = \realpath(\dirname(__DIR__, 3)) . '/Anonymization/Anonymizer';
+
+        return $config;
+    }
+
     #[\Override]
     public function getConfigTreeBuilder(): TreeBuilder
     {
