+++ src/Bridge/Symfony/DependencyInjection/DbToolsExtension.php
@@ -21,6 +21,7 @@ final class DbToolsExtension extends Extension
     {
         $configuration = $this->getConfiguration($configs, $container);
         $config = $this->processConfiguration($configuration, $configs);
+        $config = DbToolsConfiguration::appendPostConfig($config);
 
         $loader = new YamlFileLoader($container, new FileLocator(\dirname(__DIR__).'/Resources/config'));
         $loader->load('services.yaml');
