+++ tests/Unit/Bridge/Symfony/DependencyInjection/DbToolsConfigurationTest.php
@@ -25,7 +25,7 @@ class DbToolsConfigurationTest extends TestCase
     public function testConfigurationMinimal(): array
     {
         $result = $this->processYamlConfiguration(
-            __DIR__ . '/../../Resources/config/packages/db_tools_min.yaml'
+            \dirname(__DIR__, 4) . '/Resources/config/packages/db_tools_min.yaml'
         );
 
         self::assertSame(
@@ -63,7 +63,7 @@ class DbToolsConfigurationTest extends TestCase
     public function testConfigurationAlternative1(): array
     {
         $result = $this->processYamlConfiguration(
-            __DIR__ . '/../../Resources/config/packages/db_tools_alt1.yaml'
+            \dirname(__DIR__, 4) . '/Resources/config/packages/db_tools_alt1.yaml'
         );
 
         self::assertSame(
@@ -115,7 +115,7 @@ class DbToolsConfigurationTest extends TestCase
     public function testConfigurationAlternative2(): array
     {
         $result = $this->processYamlConfiguration(
-            __DIR__ . '/../../Resources/config/packages/db_tools_alt2.yaml'
+            \dirname(__DIR__, 4) . '/Resources/config/packages/db_tools_alt2.yaml'
         );
 
         self::assertSame(
